Split App and Api controllers into namespaces
