+++ app/assets/javascripts/integrate/controllers/competition.js.coffee
@@ -6,7 +6,7 @@ angular.module('integrate').controller('CompetitionController', ($scope, $http) 
     self.unsynced_entries = data.unsynced_entries
 
   @resync_entry = (entry_id) ->
-    $http.post("/entries/#{entry_id}/resync").
+    $http.post("/api/entries/#{entry_id}/resync").
       success((data, status, headers, config) ->
         self.unsynced_entries = self.unsynced_entries.filter((entry) -> entry.id != entry_id)
       ).
