+++ app/controllers/api/base_controller.rb
@@ -1,0 +1,5 @@
+module Api
+  class BaseController < ApplicationController
+    protect_from_forgery with: :null_session
+  end
+end
