+++ app/controllers/app/base_controller.rb
@@ -1,0 +1,5 @@
+module App
+  class BaseController < ApplicationController
+    protect_from_forgery with: :exception
+  end
+end
