+++ app/controllers/app/competitions_controller.rb
@@ -1,0 +1,12 @@
+module App
+  class CompetitionsController < BaseController
+    def entrant_page
+      @entry = Entry.new(competition: Competition.find(params[:competition_id]))
+    end
+
+    def show
+      # TODO: add authentication
+      @competition = Competition.find(params[:id])
+    end
+  end
+end
