+++ app/controllers/app/welcome_controller.rb
@@ -1,0 +1,7 @@
+module App
+  class WelcomeController < BaseController
+    def index
+      @competitions = Competition.all.order('created_at desc')
+    end
+  end
+end
