--- app/controllers/welcome_controller.rb
@@ -1,5 +1,0 @@
-class WelcomeController < ApplicationController
-  def index
-    @competitions = Competition.all.order('created_at desc')
-  end
-end
