+++ app/services/create_competition.rb
@@ -1,0 +1,5 @@
+class CreateCompetition
+  def call(_competition_params)
+    { success: false, errors: { id: ['not implemented yet'] } }
+  end
+end
