+++ app/views/app/competitions/show.erb
