+++ config/routes.rb
@@ -4,11 +4,14 @@ Rails.application.routes.draw do
   # See how all your routes lay out with "rake routes".
 
   # You can have the root of your site routed with "root"
-  root 'welcome#index'
-  get ':competition_id/:permalink' => 'competitions#entrant_page', constraints: { competition_id: /\d+/ }
-  resources :competitions, only: :show
-  resources :entries, only: :create do
-    post :resync, on: :member
+  root 'app/welcome#index'
+  get ':competition_id/:permalink' => 'app/competitions#entrant_page', constraints: { competition_id: /\d+/ }
+  resources :competitions, only: :show, module: :app
+
+  namespace :api do
+    resources :entries, only: :create do
+      post :resync, on: :member
+    end
   end
 
   # Example of regular route:
